Universal business card using RFID
Finding change in shops has become a universal problem in India. Many
times either the shop owner or the customer has to lose the change
because of non-availability of the change. To address this problem this
project is proposed using RFID smart card to maintain prepaid balance in
the smartcard.
Using Smart card reader available at the shop the user can pay the
change. The shop owners can collect the cash from the Universal Business
Card Service Branch in the corresponding location.